+++ Source/CompanyCommunicator/Controllers/SentNotificationsController.cs
@@ -8,10 +8,13 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Controllers
     using System;
     using System.Collections.Generic;
     using System.Linq;
+    using System.Net.Http;
     using System.Security.Claims;
+    using System.Text;
     using System.Threading.Tasks;
     using Microsoft.AspNetCore.Authorization;
     using Microsoft.AspNetCore.Mvc;
+    using Microsoft.Azure.WebJobs.Extensions.DurableTask;
     using Microsoft.Extensions.Logging;
     using Microsoft.Extensions.Options;
     using Microsoft.Graph;
@@ -27,6 +30,7 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Controllers
     using Microsoft.Teams.Apps.CompanyCommunicator.Common.Services.MicrosoftGraph;
     using Microsoft.Teams.Apps.CompanyCommunicator.Controllers.Options;
     using Microsoft.Teams.Apps.CompanyCommunicator.Models;
+    using Newtonsoft.Json;
 
     /// <summary>
     /// Controller for the sent notification data.
@@ -46,6 +50,7 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Controllers
         private readonly IAppCatalogService appCatalogService;
         private readonly IAppSettingsService appSettingsService;
         private readonly UserAppOptions userAppOptions;
+        private readonly IHttpClientFactory clientFactory;
         private readonly ILogger<SentNotificationsController> logger;
 
         /// <summary>
@@ -62,6 +67,7 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Controllers
         /// <param name="appCatalogService">App catalog service.</param>
         /// <param name="appSettingsService">App settings service.</param>
         /// <param name="userAppOptions">User app options.</param>
+        /// <param name="clientFactory">the http client factory.</param>
         /// <param name="loggerFactory">The logger factory.</param>
         public SentNotificationsController(
             INotificationDataRepository notificationDataRepository,
@@ -75,6 +81,7 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Controllers
             IAppCatalogService appCatalogService,
             IAppSettingsService appSettingsService,
             IOptions<UserAppOptions> userAppOptions,
+            IHttpClientFactory clientFactory,
             ILoggerFactory loggerFactory)
         {
             if (dataQueueMessageOptions is null)
@@ -93,6 +100,7 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Controllers
             this.appCatalogService = appCatalogService ?? throw new ArgumentNullException(nameof(appCatalogService));
             this.appSettingsService = appSettingsService ?? throw new ArgumentNullException(nameof(appSettingsService));
             this.userAppOptions = userAppOptions?.Value ?? throw new ArgumentNullException(nameof(userAppOptions));
+            this.clientFactory = clientFactory ?? throw new ArgumentNullException(nameof(clientFactory));
             this.logger = loggerFactory?.CreateLogger<SentNotificationsController>() ?? throw new ArgumentNullException(nameof(loggerFactory));
         }
 
@@ -168,6 +176,7 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Controllers
                     Succeeded = notificationEntity.Succeeded,
                     Failed = notificationEntity.Failed,
                     Unknown = this.GetUnknownCount(notificationEntity),
+                    Canceled = notificationEntity.Canceled > 0 ? notificationEntity.Canceled : (int?)null,
                     TotalMessageCount = notificationEntity.TotalMessageCount,
                     SendingStartedDate = notificationEntity.SendingStartedDate,
                     Status = notificationEntity.GetStatus(),
@@ -222,6 +231,7 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Controllers
                 Succeeded = notificationEntity.Succeeded,
                 Failed = notificationEntity.Failed,
                 Unknown = this.GetUnknownCount(notificationEntity),
+                Canceled = notificationEntity.Canceled > 0 ? notificationEntity.Canceled : (int?)null,
                 TeamNames = await this.teamDataRepository.GetTeamNamesByIdsAsync(notificationEntity.Teams),
                 RosterNames = await this.teamDataRepository.GetTeamNamesByIdsAsync(notificationEntity.Rosters),
                 GroupNames = groupNames,
@@ -234,6 +244,53 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Controllers
             };
 
             return this.Ok(result);
+        }
+
+        /// <summary>
+        /// Cancel the sent notification by id.
+        /// </summary>
+        /// <param name="id">notification id.</param>
+        /// <returns>The result of an action method.</returns>
+        [HttpPost("cancel/{id}")]
+        public async Task<IActionResult> CancelSentNotificationByIdAsync(string id)
+        {
+            _ = id ?? throw new ArgumentNullException(nameof(id));
+
+            var notificationDataEntity = await this.notificationDataRepository.GetAsync(
+                NotificationDataTableNames.SentNotificationsPartition,
+                id);
+            if (notificationDataEntity == null)
+            {
+                return this.NotFound();
+            }
+
+            var instancePayload = JsonConvert.DeserializeObject<HttpManagementPayload>(notificationDataEntity.FunctionInstancePayload);
+            var client = this.clientFactory.CreateClient();
+            var httpContent = new StringContent(string.Empty, Encoding.UTF8, "application/json");
+
+            // Update the reason of termination.
+            var terminateUri = instancePayload.TerminatePostUri.Replace("{text}", "Canceled");
+            var response = await client.PostAsync(terminateUri, httpContent);
+            if (response.StatusCode == System.Net.HttpStatusCode.Accepted ||
+                response.StatusCode == System.Net.HttpStatusCode.Gone)
+            {
+                if (!notificationDataEntity.IsCompleted())
+                {
+                    notificationDataEntity.Status = NotificationStatus.Canceling.ToString();
+                    await this.notificationDataRepository.InsertOrMergeAsync(notificationDataEntity);
+
+                    // send message to data queue
+                    var messageDelay = new TimeSpan(0, 0, 5);
+                    await this.dataQueue.SendMessageAsync(id, messageDelay);
+                    return this.Accepted();
+                }
+            }
+            else if (response.StatusCode == System.Net.HttpStatusCode.NotFound)
+            {
+                return this.NotFound();
+            }
+
+            return this.Ok();
         }
 
         private int? GetUnknownCount(NotificationDataEntity notificationEntity)